Fixed the clear all tasks. It now clears all task only for that particular user

+++ app/src/main/java/sg/edu/np/mad/madassignment1/DBHandler.java
@@ -223,9 +223,9 @@ public class DBHandler extends SQLiteOpenHelper {
         // db.delete("Tasks","TaskName=? and TaskDesc=?",new String[]{taskName,taskDesc});
     }
 
-    public void deleteAllTask(){
+    public void deleteAllTask(int userID){
         SQLiteDatabase db = this.getWritableDatabase();
-        db.execSQL("delete from "+ TASKS);
+        db.execSQL("delete from "+ TASKS + " WHERE " + COLUMN_USERID + "=" + userID);
         db.close();
     }
